BOXING VAR - ML TECHNOLOGY
FOR A BOXING MATCH - MACHINE LEARNING SYSTEM FOR AUTOMMATED COUNTER OF OPPONENT POINTS- VIRTUAL ASSISTED REFEREE (VAR) 
INSTRUCTIONS AS FOLLOWS:
It is proposed for Professional Boxing to also start using technology to determine touches. Some sort of Video Assisted Referee. My proposal is backed with Data Science, under number series and cumulative counters. Let each boxer wear a skin tight but well ventilated Full Upper Body vest {from waist to head] which is lined with some charged touch sensors to accurately count every time the opponents hit each other. In order to separate the touches from that of the ring referee, make the opponents VESTS to have Positive Charges (+) while the GLOVES have Negative Charges (-) so every time there is a hit or touch, this creates a COMPLETE electrical circuit for which a LIGHT or SOUND will be heard together with an automatic counter database. Let’s call the innovation [Boxing VAR]. This will then tally the outcomes and compare with that of the referees at end of each fight!